Clean up clippy lints (semantics-preserving)
Apply safe auto-fixable lints across config, key handling, render,
transport, and the native quiet/find/copy paths: io::Error::other,
derivable Default for Theme, nonminimal boolean simplification, redundant
casts, explicit auto-deref, map_or -> is_none_or, split_once, and
clippy::type_complexity allow on the ANSI deserializer. No behavior change.

114 unit + 2 E2E pass; release + fmt clean; clippy back under ~6 real lints.

+++ src/render.rs
@@ -1159,7 +1159,7 @@ fn color256(i: u8, colors: &Colors) -> (u8, u8, u8) {
             if v == 0 {
                 0
             } else {
-                (55 + v * 40) as u8
+                55 + v * 40
             }
         };
         (ramp(r), ramp(g), ramp(b))
